Still working out Bellman Ford

+++ src/Common/Forex/Arbitrage.cs
@@ -3,13 +3,13 @@ namespace Common.Forex
 
     public class Arbitrage
     {
-        public Arbitrage(ExchangeChain chain, decimal ratio)
+        public Arbitrage(ExchangeChain chain, double ratio)
         {
             this.Chain = chain;
             this.Ratio = ratio;
 
         }
         public ExchangeChain Chain { get; set; }
-        public decimal Ratio { get; set; }
+        public double Ratio { get; set; }
     }
 }
