+++ src/Common/Forex/Exchange.cs
@@ -4,7 +4,7 @@ namespace Common.Forex
 {
     public class Exchange : IEquatable<Exchange>
     {
-        public Exchange(Currency oldCurrency, Currency newCurrency, decimal exchangeRate = 1)
+        public Exchange(Currency oldCurrency, Currency newCurrency, double exchangeRate = 1)
         {
             this.OldCurrency = oldCurrency;
             this.NewCurrency = newCurrency;
@@ -12,7 +12,7 @@ namespace Common.Forex
         }
         public Currency OldCurrency { get; set; }
         public Currency NewCurrency { get; set; }
-        public decimal ExchangeRate { get; set; }
+        public double ExchangeRate { get; set; }
         public bool Equals(Exchange other) => this.OldCurrency == other.OldCurrency && this.NewCurrency == other.NewCurrency && this.ExchangeRate == other.ExchangeRate;
         public override int GetHashCode() => ToString().GetHashCode();
         public override string ToString() => $"{OldCurrency} to {NewCurrency} at a rate of {ExchangeRate}";
